Added support for multisampling multiple buffers.
Color buffers are no longer transitioned to the
VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL layout, when recreating the
swapchain. Instead there now is a check to see if they are input
attachments or sampled from.

Extended the example to feature multisampling in the g_buffer subpass.
Multisampling is only performed on the position and normal buffers.